Reading header from stream
I need to make this testable. Guess I'm writing header length as start
of header

+++ dbms/src/lib.rs
@@ -50,7 +50,7 @@ pub enum SqlError {
     IncompatibleVersion(u8),
 
     InvalidHeader(&'static str),
-    InvalidMessageType(u64),
+    InvalidMessageType(u8),
 
     CouldNotWriteToConnection(std::io::Error),
     CouldNotReadFromConnection(std::io::Error),
